{X00TD, X01BD}_defconfig: Enable INIT_STACK_ALL_ZERO
Init stack is secure as well usefull.
As per logs we have certain uninitialised variables.
so this will be a valid fix for them.

Note that as per comments the usage of this is controversial
But, no adverse effects were noticed on our system
So, I feel it's safe to use on our devices

Change-Id: I0d2e786237b383dfd6e4d3f83b59fdb4a62a00e6
Signed-off-by: Kunmun <kunmun.devroms@gmail.com>